requirements.txt
Flask
tensorflow-cpu
numpy
opencv-python
gunicorn